consumer/app/requirements.txt
kafka-python
pymongo
joblib
scikit-learn